+++ History.md
@@ -2,6 +2,7 @@ Unpublished
 ===========
 
   * check for valid arguments in addon methods
+  * added sync and async tests
 
 0.3.0 / Dec 13 2014
 ===================